# docs/openapi.yaml
openapi: 3.0.4
info:
  title: Needle Thread API (Public)
  version: 1.0.0
  description: |
    Public OpenAPI specification for the Thread API.
    This file intentionally excludes private/admin authentication endpoints and
    internal-only schema details.
servers:
  - url: http://localhost:5000
    description: Local development

tags:
  - name: System
  - name: Events
  - name: Images
  - name: Members

paths:
  /:
    get:
      tags: [System]
      summary: Service root
      operationId: getRoot
      responses:
        "200":
          description: Welcome message
          content:
            text/plain:
              schema:
                type: string
                example: Welcome to Needle!

  /health:
    get:
      tags: [System]
      summary: Health check
      operationId: getHealth
      responses:
        "200":
          description: Service and DB are healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthSuccessResponse"
        "500":
          description: Service unhealthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /events/{event_id}:
    get:
      tags: [Events]
      summary: Get event by id
      description: Returns a published event document for the provided id.
      operationId: getEventById
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      responses:
        "200":
          description: Event found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetEventSuccessResponse"
        "400":
          description: Invalid event id format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /events/past:
    get:
      tags: [Events]
      summary: Get recent past events
      operationId: getPastEvents
      responses:
        "200":
          description: Past events payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetPastEventsSuccessResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /events/overview:
    get:
      tags: [Events]
      summary: Get event overview
      operationId: getEventOverview
      responses:
        "200":
          description: Overview payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetEventOverviewSuccessResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /events/semester/{semester_id}:
    get:
      tags: [Events]
      summary: Get events by semester code
      operationId: getSemesterEvents
      parameters:
        - name: semester_id
          in: path
          required: true
          schema:
            type: string
            pattern: "^[SAF][0-9]{2}$"
            example: F26
      responses:
        "200":
          description: Semester events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetSemesterSuccessResponse"
        "400":
          description: Invalid semester code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /images/{image_id}:
    get:
      tags: [Images]
      summary: Get image by id
      operationId: getImageById
      parameters:
        - name: image_id
          in: path
          required: true
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      responses:
        "200":
          description: Image found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetImageSuccessResponse"
        "400":
          description: Invalid image id format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Image not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /members/{id}:
    get:
      tags: [Members]
      summary: Get member by id
      operationId: getMemberById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      responses:
        "200":
          description: Member found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetMemberSuccessResponse"
        "400":
          description: Invalid member id format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Member not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /members/{year}:
    get:
      tags: [Members]
      summary: Get members for a two-digit year
      operationId: getMembersByYear
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: string
            pattern: "^[0-9]{2}$"
            example: "26"
      responses:
        "200":
          description: Members list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetMembersSuccessResponse"
        "400":
          description: Invalid year parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      additionalProperties: true

    HealthSuccessResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [state]
          properties:
            state:
              type: string
              example: healthy

    EventType:
      type: string
      enum: [shoot, internal, external]

    AdditionalPersonnelItem:
      type: object
      required: [member_id, role]
      properties:
        member_id:
          type: string
        role:
          type: string

    PublicEvent:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
        date:
          type: string
          format: date-time
        location:
          type: string
        type:
          $ref: "#/components/schemas/EventType"
        image_path:
          type: string
        blurb:
          type: string
        published:
          type: boolean
        photographer_ids:
          type: array
          items:
            type: string
        creative_director_ids:
          type: array
          items:
            type: string
        model_ids:
          type: array
          items:
            type: string
        additional_personnel:
          type: array
          items:
            $ref: "#/components/schemas/AdditionalPersonnelItem"
      additionalProperties: true

    PublicImage:
      type: object
      properties:
        id:
          type: string
        event_id:
          type: string
        date:
          type: string
          format: date-time
        path:
          type: string
        caption:
          type: string
        published:
          type: boolean
        width:
          type: integer
        height:
          type: integer
        photographer_id:
          type: string
        creative_director_id:
          type: string
        model_ids:
          type: array
          items:
            type: string
        additional_personnel:
          type: array
          items:
            $ref: "#/components/schemas/AdditionalPersonnelItem"
      additionalProperties: true

    PublicMember:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        display_name:
          type: string
        pronouns:
          type: string
        role:
          type: string
        email:
          type: string
          format: email
        links:
          type: object
          additionalProperties:
            type: string
        profile_photo_path:
          type: string
        blurb:
          type: string
        start_year:
          type: integer
        end_year:
          type: integer
          nullable: true
        visibility:
          type: string
        active:
          type: boolean
      additionalProperties: true

    PastEventCard:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        date:
          type: string
          format: date-time
        location:
          type: string
        type:
          $ref: "#/components/schemas/EventType"
        image_path:
          type: string
      additionalProperties: true

    GetEventSuccessResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          allOf:
            - $ref: "#/components/schemas/PublicEvent"

    GetImageSuccessResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/PublicImage"

    GetMemberSuccessResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/PublicMember"

    GetPastEventsSuccessResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [past_events]
          properties:
            past_events:
              type: array
              items:
                $ref: "#/components/schemas/PastEventCard"

    GetEventOverviewSuccessResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [past, upcoming]
          properties:
            past:
              type: array
              items:
                $ref: "#/components/schemas/PublicEvent"
            upcoming:
              type: array
              items:
                $ref: "#/components/schemas/PublicEvent"

    GetSemesterSuccessResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [semester_events]
          properties:
            semester_events:
              type: array
              items:
                $ref: "#/components/schemas/PublicEvent"

    MemberListItem:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        role:
          type: string
        blurb:
          type: string
        start_year:
          type: integer
        end_year:
          type: integer
          nullable: true
      additionalProperties: true

    GetMembersSuccessResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MemberListItem"